pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
]
